// src/utils/monica/types.ts
export interface AlchemicalState {
  // Elemental
  fire: number;
  water: number;
  air: number;
  earth: number;
  // Alchemical
  spirit: number;
  essence: number;
  matter: number;
  substance: number;
}

export interface MonicaMetrics {
  heat: number;
  entropy: number;
  reactivity: number;
  gregsEnergy: number;
  kAlchm: number;
  monicaConstant: number;
}